+++ spec/unit/validator_spec.rb
@@ -998,7 +998,7 @@ module Pod
           target.deployment_target.should.be.nil
           target.platform_name.should == :ios
           target.build_configurations.each do |c|
-            File.basename(c.build_settings['INFOPLIST_FILE']).should == 'App-Info.plist'
+            c.build_settings['INFOPLIST_FILE'].should == '$(SRCROOT)/App/App-Info.plist'
           end
           Xcodeproj::Project.schemes(project.path).should == %w(App)
         end